Add status line to README
As mentioned in my GA Coursework repo, the individual project repos are to have
a note once the status of the project is complete is to be place in the README
for that project. Having a place to put it, and showing the status as not yet
complete, wouldn't hurt.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>